requirements.txt
3-1==1.0.0
appnope==0.1.0
argon2-cffi==20.1.0
asgiref==3.3.1
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
cachetools==4.1.1
certifi==2020.4.5.2
cffi==1.14.1
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.6
django-allauth==0.42.0
django-currentuser==0.5.2
django-custom-user==0.7
django-debug-toolbar==2.2
django-etc==1.2.0
django-extensions==3.0.4
django-model-utils==4.0.0
django-notifications-hq==1.6.0
django-ordered-model==3.4.1
django-siteflags==1.2.0
django-storages==1.9.1
django-vote==2.2.0
djangorestframework==3.11.0
entrypoints==0.3
gensim==3.8.3
google-api-core==1.24.1
google-auth==1.24.0
google-cloud-core==1.3.0
google-cloud-storage==1.29.0
google-cloud-translate==3.0.2
google-cloud-vision==2.0.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
googletrans==3.1.0a0
grpcio==1.34.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.9
importlib-metadata==1.7.0
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jsonfield==3.1.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
libcst==0.3.16
MarkupSafe==1.1.1
mistune==0.8.4
mypy-extensions==0.4.3
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.1.1
numpy==1.19.4
oauthlib==3.1.0
packaging==20.4
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pi==0.1.2
pickleshare==0.7.5
Pillow==7.1.2
prometheus-client==0.8.0
prompt-toolkit==3.0.5
proto-plus==1.13.0
protobuf==3.12.2
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.5
QtPy==1.9.0
requests==2.23.0
requests-oauthlib==1.3.0
rfc3986==1.4.0
rsa==4.6
scipy==1.6.0
Send2Trash==1.5.0
six==1.15.0
smart-open==4.1.0
sniffio==1.2.0
sqlparse==0.3.1
swapper==1.1.2.post1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
tqdm==4.55.1
traitlets==4.3.3
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.9
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.1.0
